requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
black==24.2.0
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.3
crispy-bootstrap4==2023.1
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.60.1
django-allauth-ui==0.2.2
django-appconf==1.0.6
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-js-asset==2.2.0
django-widget-tweaks==1.5.0
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy==1.8.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.0.2
pytest-django==4.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
questionary==2.0.1
rcssmin==1.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rich-click==1.7.3
rjsmin==1.2.2
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.4
types-bleach==6.1.0.20240222
typing_extensions==4.10.0
Unidecode==1.3.8
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.6.0
pillow==10.2.0